requirements.txt
annotated-types==0.7.0
anyio==4.6.2
certifi==2025.6.15
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.13
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.1
openai==1.90.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.44.0
tqdm==4.67.1
typing_extensions==4.13.2
uvicorn==0.33.0
transformers==4.46.3
torch==2.2.2
accelerate>=0.26.0
bitsandbytes==0.42.0
python-jose==3.4.0
tiktoken==0.7.0
langchain
langchain-core>=0.1.70
langchain-community
langchain-openai
langchain-chroma
langgraph
langchain-huggingface
tavily-python
pymongo[srv]
faiss-cpu==1.12.0
PyMuPDF
langchain-tavily==0.2.12
mysql-connector-python
tabulate
sentence-transformers
langchain-huggingface
hf_xet
